+++ index.html
@@ -10,17 +10,16 @@
 </head>
 
 <body>
-  <div class="container">
-    <ul class="comments" id="list" data>
-<!-- Рендеринг в JS -->
-    </ul>
+  <div class="container"> 
+  <div id="app"></div>
+   <!-- <ul class="comments" id="list" data> 
+ Рендеринг в JS
+    </ul> -->
     <section id="loaderComments" class="loader -display-none">
       <h4 class="loader-text" id="loaderText">Комментарии загружаются...</h4>
       <div class="loader__gif">
         <img src="./img/Spinner-1s-100px.svg" alt="loader">
       </div>
-    </section>
-
     </section>
     <div class="add-form">
       <input id="name-input" type="text" class="add-form-name" placeholder="Введите ваше имя" />
@@ -30,10 +29,22 @@
         <button class="add-form-button-delete" id="add-button-delete">Удалить комментарий</button>
         <button class="add-form-button" id="add-button">Написать</button>
       </div>
+    </div>  
+    <!-- Рендер с JS -->
+    <!-- Форма входа  -->
+     <!-- <div class="add-form-authorization">
+      <h3 class="form-text">Форма авторизации</h3>
+      <input id="login-input" type="text" class="add-form-login" placeholder="Введите логин" />
+      <input id="password-input" type="password" class="add-form-password" placeholder="Введите пароль" />  -->
+      <!-- Флрма регистрации -->
+    
+      <!-- <div class="add-form-row">
+        <button class="add-form-button" id="login-button">Войти</button>
+      </div>
     </div>
-  </div>
-  </div>
-  <!-- <script src="./js/script.js"></script> -->
+  </div> 
+  </div> -->
+  
   <script type="module" src="js/script.js"></script>
 </body>
