+++ js/script.js
@@ -1,54 +1,57 @@
-import {comments, addComments, fetchGetAndRenderComments} from "./requests.js"
-import {ButtonTouch} from "./likes.js"
 
-export const likeButtons = document.querySelectorAll(".like-button");
-export const buttonElement = document.getElementById("add-button");
-export const nameElement = document.getElementById("name-input");
-export const commentsElement = document.getElementById("comments-input");
-export const listElement = document.getElementById("list");
+
+const likeButtons = document.querySelectorAll(".like-button");
+const buttonElement = document.getElementById("add-button");
+const nameElement = document.getElementById("name-input");
+const commentsElement = document.getElementById("comments-input");
+const listElement = document.getElementById("list");
 const commentListElement = document.getElementById("comment-list");
-export const mainForm = document.querySelector(".add-form");
+const mainForm = document.querySelector(".add-form");
 const loaderText = document.getElementById("loaderText");
-export const loaderCommentsElement = document.getElementById("loaderComments");
-buttonElement.disabled = true
+const loaderCommentsElement = document.getElementById("loaderComments");
+let token = "Bearer cwboc4d0co6gckd8b8cocwdg6g39k3c03cg3d83cw3k39c3b43cg3b43bw3b43ck3co3bc"
+buttonElement.disabled = true;
 
 // GET запрос 
-// const fetchGetAndRenderComments = () => {
-//   return fetch("https://webdev-hw-api.vercel.app/api/v1/timur-ramazanov/comments", {
-//     method: "GET"
-//   }).then((response) => {
-//     if (response.status === 200) {
-//       return response.json();
-//     } else {
-//       throw new Error("Что то пошло не так")
-//     }
-//   }).then((responseData) => {
-//     const options = {
-//       year: '2-digit',
-//       month: 'numeric',
-//       day: 'numeric',
-//       timezone: 'UTC',
-//       hour: 'numeric',
-//       minute: '2-digit'
-//     };
-//     const appComments = responseData.comments.map((comment) => {
-//       return {
-//         name: comment.author.name,
-//         date: new Date(comment.date).toLocaleDateString("ru-RU", options).replace(',', ' '),
-//         text: comment.text,
-//         likeCount: comment.likes,
-//         liked: false,
-//       };
-//     });
-//     comments = appComments;
-//     loaderCommentsElement.classList.add('-display-none')
-//     renderComments();
-//   }).catch((error) => {
-//     alert('Проверьте подключение к интернету');
-//     console.warn(error);
-//   })
-// }
-// let comments = [];
+const fetchGetAndRenderComments = () => {
+  return fetch("https://webdev-hw-api.vercel.app/api/v2/timur-ramazanov/comments", {
+    method: "GET",
+    headers: {
+      Authorization: token,
+    },
+  }).then((response) => {
+    if (response.status === 200) {
+      return response.json();
+    } else {
+      throw new Error("Что то пошло не так")
+    }
+  }).then((responseData) => {
+    const options = {
+      year: '2-digit',
+      month: 'numeric',
+      day: 'numeric',
+      timezone: 'UTC',
+      hour: 'numeric',
+      minute: '2-digit'
+    };
+    const appComments = responseData.comments.map((comment) => {
+      return {
+        name: comment.author.name,
+        date: new Date(comment.date).toLocaleDateString("ru-RU", options).replace(',', ' '),
+        text: comment.text,
+        likeCount: comment.likes,
+        liked: false,
+      };
+    });
+    comments = appComments;
+    loaderCommentsElement.classList.add('-display-none')
+    renderComments();
+  }).catch((error) => {
+    alert('Проверьте подключение к интернету');
+    console.warn(error);
+  })
+}
+let comments = [];
 
 
 
@@ -65,35 +68,34 @@ const buttonBlock = () => {
   });
 }
 // // Кнопка лайка
-// const ButtonTouch = () => {
-//   const likeButtons = document.querySelectorAll(".like-button");
-//   for (const likeButton of likeButtons) {
-//     likeButton.addEventListener("click", () => {
-//       console.log(11);
-//       const index = likeButton.dataset.index;
+const ButtonTouch = () => {
+  const likeButtons = document.querySelectorAll(".like-button");
+  for (const likeButton of likeButtons) {
+    likeButton.addEventListener("click", () => {
+      const index = likeButton.dataset.index;
 
-//       if (comments[index].liked === false) {
-//         comments[index].liked = true;
-//         comments[index].likeCount += 1;
+      if (comments[index].liked === false) {
+        comments[index].liked = true;
+        comments[index].likeCount += 1;
 
-//       } else {
-//         comments[index].liked = false;
-//         comments[index].likeCount -= 1;
-//       }
-//       delay(2000).then(() => {
-//         renderComments();
+      } else {
+        comments[index].liked = false;
+        comments[index].likeCount -= 1;
+      }
+      delay(2000).then(() => {
+        renderComments();
 
-//       })
-//     })
-//   };
-// }
-// function delay(interval = 300) {
-//   return new Promise((resolve) => {
-//     setTimeout(() => {
-//       resolve();
-//     }, interval);
-//   });
-// }
+      })
+    })
+  };
+}
+function delay(interval = 300) {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, interval);
+  });
+}
 // Цитирует выбранный комментарий
 const initTouchComment = () => {
   const touchComments = listElement.querySelectorAll(".comment-text");
@@ -117,151 +119,134 @@ const buttonDelete = () => {
 buttonDelete();
 
 // Повторный POST запрос при ошибке 500
-// function repeatedRequestComment() {
-//   mainForm.classList.add('-display-none');
-//   loaderCommentsElement.classList.remove('-display-none');
-//   // Передаем комментарий
-//   fetch("https://webdev-hw-api.vercel.app/api/v1/timur-ramazanov/comments", {
-//     method: "POST",
-//     body: JSON.stringify({
-//       name: nameElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
-//       date: new Date,
-//       text: commentsElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
-//       likeCount: 0,
-//       liked: false,
-//       forceError: true,
-//     })
-//   }).then((response) => {
-//     if (response.status === 201) {
-//       return response.json();
-//     } else if (response.status === 400) {
-//       throw new Error("Слишком коротко")
-//     } else {
-//       throw new Error("Сервер сломался")
-//     }
-//   }).then((responseData) => {
-//     comments = responseData;
-//   }).then(() => {
-//     return fetchGetAndRenderComments();
+function repeatedRequestComment() {
+  mainForm.classList.add('-display-none');
+  loaderCommentsElement.classList.remove('-display-none');
+  // Передаем комментарий
+  fetch("https://webdev-hw-api.vercel.app/api/v2/timur-ramazanov/comments", {
+    method: "POST",
+    body: JSON.stringify({
+      name: nameElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
+      date: new Date,
+      text: commentsElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
+      likeCount: 0,
+      liked: false,
+      forceError: true,
+    }),
+    headers: {
+      Authorization: token,
+    },
+  }).then((response) => {
+    if (response.status === 201) {
+      return response.json();
+    } else if (response.status === 400) {
+      throw new Error("Слишком коротко")
+    } else {
+      throw new Error("Сервер сломался")
+    }
+  }).then((responseData) => {
+    comments = responseData;
+  }).then(() => {
+    return fetchGetAndRenderComments();
 
-//   }).then((data) => {
-//     mainForm.classList.remove('-display-none');
-//     loaderCommentsElement.classList.add('-display-none');
+  }).then((data) => {
+    mainForm.classList.remove('-display-none');
+    loaderCommentsElement.classList.add('-display-none');
 
-//     nameElement.value = '';
-//     commentsElement.value = '';
-//     buttonElement.disabled = true;
-//   }).catch((error) => {
-//     if (error.message === 'Слишком коротко') {
-//       alert("name и text должены содержать хотя бы 3 символа");
-//       mainForm.classList.remove('-display-none');
-//       loaderCommentsElement.classList.add('-display-none');
+    nameElement.value = '';
+    commentsElement.value = '';
+    buttonElement.disabled = true;
+  }).catch((error) => {
+    if (error.message === 'Слишком коротко') {
+      alert("name и text должены содержать хотя бы 3 символа");
+      mainForm.classList.remove('-display-none');
+      loaderCommentsElement.classList.add('-display-none');
 
-//     } if (error.message === 'Сервер сломался') {
-//       repeatedRequestComment();
-//       console.warn(error);
-//     }
-
-
-//   });
-//   renderComments();
-// };
+    } if (error.message === 'Сервер сломался') {
+      repeatedRequestComment();
+      console.warn(error);
+    }
 
 
-
-// const addComments = () => {
-//   buttonElement.addEventListener("click", () => {
-
-//     nameElement.classList.remove("error");
-//     commentsElement.classList.remove("error");
-
-//     nameElement.classList.remove("error");
-//     commentsElement.classList.remove("error");
-
-//     if (nameElement.value === '') {
-//       nameElement.classList.add("error");
-//       return
-
-//     } else if (commentsElement.value === '') {
-//       commentsElement.classList.add("error");
-//       return
-//     }
-//     mainForm.classList.add('-display-none');
-//     loaderCommentsElement.classList.remove('-display-none');
-
-//     // Передаем комментарий
-//     fetch("https://webdev-hw-api.vercel.app/api/v1/timur-ramazanov/comments", {
-//       method: "POST",
-//       body: JSON.stringify({
-//         name: nameElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
-//         date: new Date,
-//         text: commentsElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
-//         likeCount: 0,
-//         liked: false,
-//         forceError: true,
-//       })
-//     }).then((response) => {
-//       if (response.status === 201) {
-//         return response.json();
-//       } else if (response.status === 400) {
-//         throw new Error("Слишком коротко")
-//       } else {
-//         throw new Error("Сервер сломался")
-//       }
-//     }).then((responseData) => {
-//       comments = responseData;
-//     }).then(() => {
-//       return fetchGetAndRenderComments();
-
-//     }).then((data) => {
-//       mainForm.classList.remove('-display-none');
-//       loaderCommentsElement.classList.add('-display-none');
-
-//       nameElement.value = '';
-//       commentsElement.value = '';
-//       buttonElement.disabled = true;
-//     }).catch((error) => {
-//       if (error.message === 'Слишком коротко') {
-//         alert("name и text должены содержать хотя бы 3 символа");
-//         mainForm.classList.remove('-display-none');
-//         loaderCommentsElement.classList.add('-display-none');
-
-//       } else if (error.message === 'Сервер сломался') {
-//         console.warn(error);
-//         repeatedRequestComment();
-//       }
-
-
-//     });
-//     renderComments();
-//   })
-// };
+  });
+  renderComments();
+};
 
 
 
+const addComments = () => {
+  buttonElement.addEventListener("click", () => {
 
-// Пока не разобрался... 
-// const editButtonTouch = () =>{
+    nameElement.classList.remove("error");
+    commentsElement.classList.remove("error");
 
-//   const editButtons = document.querySelectorAll(".edit-button");
-//   for(editButton of editButtons){
+    nameElement.classList.remove("error");
+    commentsElement.classList.remove("error");
 
-//     editButton.addEventListener("click", () =>{
-//       const index = editButton.dataset.index;
-//       if(comments[index].isEdit === false){
-//         comments[index].isEdit = true;
+    if (nameElement.value === '') {
+      nameElement.classList.add("error");
+      return
 
-//       }else{
-//         comments[index].isEdit = false;
-//       }
-//     });
+    } else if (commentsElement.value === '') {
+      commentsElement.classList.add("error");
+      return
+    }
+    mainForm.classList.add('-display-none');
+    loaderCommentsElement.classList.remove('-display-none');
 
-//   }
-// };
+    // Передаем комментарий
+    fetch("https://webdev-hw-api.vercel.app/api/v2/timur-ramazanov/comments", {
+      method: "POST",
+      body: JSON.stringify({
+        name: nameElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
+        date: new Date,
+        text: commentsElement.value.replaceAll("<", "&lt;").replaceAll(">", "&gt;"),
+        likeCount: 0,
+        liked: false,
+        forceError: true,
+      }),
+      headers: {
+        Authorization: token,
+      },
+    }).then((response) => {
+      if (response.status === 201) {
+        return response.json();
+      } else if (response.status === 400) {
+        throw new Error("Слишком коротко")
+      } else {
+        throw new Error("Сервер сломался")
+      }
+    }).then((responseData) => {
+      comments = responseData;
+    }).then(() => {
+      return fetchGetAndRenderComments();
+
+    }).then((data) => {
+      mainForm.classList.remove('-display-none');
+      loaderCommentsElement.classList.add('-display-none');
+
+      nameElement.value = '';
+      commentsElement.value = '';
+      buttonElement.disabled = true;
+    }).catch((error) => {
+      if (error.message === 'Слишком коротко') {
+        alert("name и text должены содержать хотя бы 3 символа");
+        mainForm.classList.remove('-display-none');
+        loaderCommentsElement.classList.add('-display-none');
+
+      } else if (error.message === 'Сервер сломался') {
+        console.warn(error);
+        repeatedRequestComment();
+      }
 
 
+    });
+    renderComments();
+  })
+};
 
-export const renderComments = () => {
+const renderComments = (buttonDelete) => {
+  const appEl = document.getElementById("app");
   const commentsHtml = comments.map((comment, index) => {
     return `<li class="comment" id = "comment-list" data-index="${index}">
     <div class="comment-header">
@@ -269,15 +254,13 @@ export const renderComments = () => {
       <div>${comment.date} </div>
     </div>
     <div class="comment-body">
-    ${comment.isEdit
-        ? `<textarea class="area-text">${comment.text}</textarea>`
-        : `<div class="comment-text"data-index = "${index}">`}
+    
+        <div class="comment-text"data-index = "${index}">
         ${comment.text} 
       </div>
     </div>
     <div class="comment-footer">
       <div class="likes">
-      <button class = "edit-button" data-index ="${index}">Редактировать</button>
         <span class="likes-counter">${comment.likeCount}</span>
         <button data-index="${index}" class="${comment.liked ? 'like-button -loading-like -active-like' : 'like-button'}"></button>
         
@@ -286,14 +269,17 @@ export const renderComments = () => {
   </li>`
   })
     .join('');
-  listElement.innerHTML = commentsHtml
-  ButtonTouch();
-  initTouchComment();
-  buttonBlock();
+  const appHtml = `  
+   <ul class="comments" id="list" data>
+<!-- Рендеринг в JS -->
+${commentsHtml}
+ </ul> `
+  appEl.innerHTML = appHtml;
+  const buttonDeleteElement = document.getElementById("add-button-delete");
+
+
 };
-addComments();
 fetchGetAndRenderComments();
-renderComments();
 
 mainForm.addEventListener("keyup", (e) => {
   if (e.code === 'Enter' || e.code === 'NumpadEnter') {
@@ -301,4 +287,3 @@ mainForm.addEventListener("keyup", (e) => {
 
   }
 });
-
